[config] Update project settings to modern standards
Update project and gradle settings, and tool versions, for the project to be directly buildable in Android Studio Chipmunk.

This only brings the project to a buildable state.
This doesn't modernize the implementation and obsolete / deprecated tools, libraries and APIs are left as is as much as possible.
For this reason, the app builds and runs from Android Studio Chipmunk, but location permission must be manually granted by the user on the device after installation as the code isn't updated to handle the "new" permission API.

The README is updated to reflect this, and better inform on this app relevance or abscence thereof, as well as caveats when running on a modern Android OS version.