[doc] Move README screenshots to another branch
Documentation assets like screenshots for the README have been moved from a `gh-pages` branch, which was misleading since no GitHub Pages is set up for this project yet, to a dedicated `assets` branch, which branches out at the root of the project history to separate it from the source code history as they should be independent.

This is basically the idea presented in [this gist](https://gist.github.com/joncardasis/e6494afd538a400722545163eb2e1fa5), but without using orphaned branches as they are not listed as regular branches and are more for single-shot publishing (as illustrated in the [official documentation](https://git-scm.com/docs/git-checkout#Documentation/git-checkout.txt---orphanltnew-branchgt)).